Can't unset filter in old templates
When using the old templates with the old style to filter you can't use the "Show all" Link to remove the filter already set.

The underlying problem is the introduction of the QueryAlias "f" for filtering, which is not set in the getFacetProperties function. By setting also the new short version to the correct value it is also possible to remove a filter when there is only one left.